fix: add validation for reward ID in DeleteRewardUseCase

+++ core/application/use-cases/rewards/commands/delete-reward.use-case.ts
@@ -1,4 +1,5 @@
 import { DeleteReward } from "@/core/application/interfaces/rewards";
+import { RewardNotFoundError } from "@/core/domain/rewards/errors";
 import { RewardRepository } from "@/core/domain/rewards/reward.repository.interface";
 
 export interface DeleteRewardDep {
@@ -13,7 +14,11 @@ export class DeleteRewardUseCase implements DeleteReward {
   }
 
   async execute(id: number): Promise<void> {
-    await this.rewardRepo.getById(id);
+    // TODO: Replace this generic error with a custom error. 
+    if (isNaN(id)) throw new Error("Invalid id");
+
+    const reward = await this.rewardRepo.getById(id);
+    if (!reward) throw new RewardNotFoundError();
     return await this.rewardRepo.delete(id);
   }
 }
